requirements.txt
boto3==1.4.6
botocore==1.6.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-dotenv==1.4.1
django-filter==1.0.4
django-oauth-toolkit==1.0.0
django-storages==1.6.5
djangorestframework==3.6.4
docutils==0.14
get==0.0.21
idna==2.6
jmespath==0.9.3
Markdown==2.6.9
oauthlib==2.0.2
post==0.0.13
psycopg2==2.7.3
public==0.0.38
python-dateutil==2.6.1
pytz==2017.2
query-string==0.0.12
requests==2.18.4
s3transfer==0.1.10
setupfiles==0.0.50
six==1.10.0
urllib3==1.22
